Use better relatime datetime strings in Media
I don't believe Home Assistant has anything totally useful here, but I
could be missing it. This adds some relative timestamp formatting for
the media captions in the Media Browser.

* Within 1 day: "N <units> ago"
* 1-7 days: "<Day-of-week>, <time>"
* 7-365 days: "<date> <time>"
* > 365 days: "<date>"

If I can find something else in HA that gives this kind of relative &
dynamic date format, I will switch to that. For now, the only thing I
could find was `dt_util.get_age()`, which is what I'm using for the
"ago" formating within 1 day. But it isn't localized (unit strings are
hard-coded, in English), and doesn't add "ago".

I see HA frontend does display "N <unit> ago", so it's possible that
this formatting is only available in the frontend, but I don't believe
I can access that from the media_source.