Filter Recent Visitor media by Feeder
Collections and Media are tied to the logged in user, but if the user
has access to multiple Feeders, the Recent Visitor entity should only
surface the visitors that came from the related Feeder.

The Collection contains a `feederName` field only (not `feederId`), and
it's unclear if the name will update as the Feeder is renamed.

Comparing by feederName is not as safe as I'd like it to be (2 Feeders
could have the same name, among other pitfalls), but it's the only
option we have.

When the visitor is handled as a result of a postcard, we are already
able to filter by feederId in this case. This applies only to looking up
the recent visitor by scanning the account's collections.

Fixes #31