Remove Feeder from Media browser
We actually cannot easily distinguish which collections and media are
from which Feeder, without fetching data for every single media_id.

Since the media collections are actually tied to the login account, it
doesn't really make sense to add Feeder to the media browser, and just
makes it harder to filter them the way you'd expect.