Handle missing collection in recent visitor check
If a photo is updated ("Wrong bird") and moved to a new collection, and
that was the last photo in the collection, then the collection will be
removed.

However, the feed item that referenced that photo and species will still
be in the Feed, even though it references something that doesn't exist.
In that case, the .collection will be explicitly set to `null`, so we
cannot use that to look up the species name.