Refresh collections in the media browser
In an attempt to avoid reloading data if we don't need to, we were
trying to use the cached collections. However, that cache might be very
sparse, which would cause broken results in the media browser.

This refreshes the collections when you enter the media browser for a
config entry. However, when you then enter one of those collections, if
we already have that collection in cache, it won't have to be reloaded
again.